# tests/PacioliBank.Contract.Tests/openapi.v1.approved.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "PacioliBank Ledger API",
    "description": "Livro-raz\u00E3o de contas correntes. Contrato descrito na EF \u00A78; valores monet\u00E1rios como n\u00FAmero decimal exato, na escala da moeda, instantes em ISO 8601 UTC.",
    "version": "1.0.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "tags": [
          "PacioliBank.Api"
        ],
        "operationId": "Vivo",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": [
          "PacioliBank.Api"
        ],
        "operationId": "Pronto",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/transfers": {
      "post": {
        "tags": [
          "Ledger"
        ],
        "summary": "Transfere entre duas contas, numa \u00FAnica transa\u00E7\u00E3o (RF-012)",
        "operationId": "Transferir",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/TransferBody"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountId}/credits": {
      "post": {
        "tags": [
          "Ledger"
        ],
        "summary": "Registra cr\u00E9dito (RF-001)",
        "operationId": "RegistrarCredito",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PostingBody"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountId}/debits": {
      "post": {
        "tags": [
          "Ledger"
        ],
        "summary": "Registra d\u00E9bito; recusa se a posi\u00E7\u00E3o ficaria negativa (RF-002)",
        "operationId": "RegistrarDebito",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PostingBody"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountId}/entries/{entryId}/reversals": {
      "post": {
        "tags": [
          "Ledger"
        ],
        "summary": "Estorna um lan\u00E7amento por compensa\u00E7\u00E3o (RF-007)",
        "operationId": "EstornarLancamento",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ReversalBody"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountId}/balance": {
      "get": {
        "tags": [
          "Ledger"
        ],
        "summary": "Posi\u00E7\u00E3o corrente ou em instante passado (RF-003, RF-004)",
        "operationId": "ConsultarPosicao",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asOf",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountId}/entries": {
      "get": {
        "tags": [
          "Ledger"
        ],
        "summary": "Extrato paginado por cursor (RF-005)",
        "operationId": "ConsultarExtrato",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatementResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BalanceResponse": {
        "required": [
          "accountId",
          "balance",
          "currency",
          "asOf",
          "computedAtSequence",
          "computedFrom",
          "entriesReplayed"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "balance": {
            "type": "number",
            "format": "decimal"
          },
          "currency": {
            "type": "string"
          },
          "asOf": {
            "type": "string"
          },
          "computedAtSequence": {
            "type": "integer",
            "format": "int64"
          },
          "computedFrom": {
            "type": "string"
          },
          "entriesReplayed": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PostingBody": {
        "type": "object",
        "properties": {
          "amount": {
            "type": [
              "null",
              "number"
            ],
            "format": "decimal"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "occurredAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "PostingResponse": {
        "required": [
          "entryId",
          "accountId",
          "sequence",
          "direction",
          "amount",
          "currency",
          "occurredAt",
          "recordedAt",
          "balanceAfter"
        ],
        "type": "object",
        "properties": {
          "entryId": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          },
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "direction": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "currency": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string"
          },
          "recordedAt": {
            "type": "string"
          },
          "balanceAfter": {
            "type": "number",
            "format": "decimal"
          },
          "reversalOf": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ReversalBody": {
        "type": "object",
        "properties": {
          "occurredAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "StatementEntryResponse": {
        "required": [
          "entryId",
          "sequence",
          "direction",
          "amount",
          "currency",
          "occurredAt",
          "recordedAt",
          "balanceAfter"
        ],
        "type": "object",
        "properties": {
          "entryId": {
            "type": "string",
            "format": "uuid"
          },
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "direction": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "currency": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string"
          },
          "recordedAt": {
            "type": "string"
          },
          "balanceAfter": {
            "type": "number",
            "format": "decimal"
          },
          "reversalOf": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "StatementResponse": {
        "required": [
          "accountId",
          "entries"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatementEntryResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "TransferBody": {
        "type": "object",
        "properties": {
          "sourceAccountId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "destinationAccountId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "amount": {
            "type": [
              "null",
              "number"
            ],
            "format": "decimal"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "occurredAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "TransferCreditLeg": {
        "required": [
          "entryId"
        ],
        "type": "object",
        "properties": {
          "entryId": {
            "type": "string"
          }
        }
      },
      "TransferDebitLeg": {
        "required": [
          "entryId",
          "sequence",
          "balanceAfter"
        ],
        "type": "object",
        "properties": {
          "entryId": {
            "type": "string"
          },
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "balanceAfter": {
            "type": "number",
            "format": "decimal"
          }
        }
      },
      "TransferResponse": {
        "required": [
          "transferId",
          "sourceAccountId",
          "destinationAccountId",
          "amount",
          "currency",
          "occurredAt",
          "recordedAt",
          "debit",
          "credit"
        ],
        "type": "object",
        "properties": {
          "transferId": {
            "type": "string"
          },
          "sourceAccountId": {
            "type": "string"
          },
          "destinationAccountId": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "currency": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string"
          },
          "recordedAt": {
            "type": "string"
          },
          "debit": {
            "$ref": "#/components/schemas/TransferDebitLeg"
          },
          "credit": {
            "$ref": "#/components/schemas/TransferCreditLeg"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "PacioliBank.Api"
    },
    {
      "name": "Ledger"
    }
  ]
}